refactor: move filetree package
